# openapi/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Proof Pass API
  version: 0.1.0
  license:
    { name: Apache 2.0, url: https://www.apache.org/licenses/LICENSE-2.0.html }
servers:
  - url: http://localhost:3000/v1
    description: Localhost development server
  - url: https://api.proofpass.io/v1
    description: Production server
paths:
...
  /user/request-verification-code:
    post:
      summary: Request an email verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserEmailVerificationRequest"
      responses:
        "200":
          description: Verification code sent successfully
  /user/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserEmailVerificationRequest:
      type: object
      properties:
        email:
          type: string
    UserLogin:
      type: object
      properties:
        email:
          type: string
        code:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
...